[GEOS-11845] Upgrade GeoServer and GeoWebCache Build to Java 17
* Upgrade build for Java 17 as base Java version
* Configure maven-enforcer-plugin to require Java 17
* Upgrade github actions workflows to use Java 17.
* Centralize maven plugin versions in the root pom's pluginManagement section.
* Reorganize root pom properties into sections
* Update docs for Java 17
* Plugin version upgrades:
  * jacoco 0.8.7 -> 0.8.13
  * maven-checkstyle-plugin 3.1.1 -> 3.6.0
  * maven-compiler-plugin 3.10.1 -> 3.14.0
  * maven-javadoc-plugin 2.10.3 -> 3.11.2
  * maven-jar-plugin 2.4 -> 3.0.2
* Upgrade Jetty 9.x to 10.0.25, last version supporting javax.servlet
* Configure Eclipse to use Java 17 for maven-eclipse-plugin
* Parameter encoding of maven-checkstyle-plugin changed to inputEncoding in 3.2
* Fix compilation errors when compiling with -Werror and Java 17. This
  is independent of the pmd and errorprone checks. Thackling them first.
  * non-varargs call of varargs method with inexact argument type for last parameter;
     * Of special interest, upgrading GridCoverageReader implementations to
       delcare varargs respecting the inherited interfaces, and fix calls
       to read(null) with read()
  * deprecated item is not annotated with @Deprecated
    * Required upgrading javacc-maven-plugin:2.3 -> 3.1.1, changing the
      depregated goal `jjtree` to `jjtree-javacc`, and updating
      wcs1_1/src/main/java/org/geoserver/wcs/kvp/rangesubset/rangeset.jjt
      to change JDK_VERSION = "1.5" to "17" and the encoding to UTF-8
  * unchecked call to put(K,V) as a member of the raw type java.util.Map
  * java.lang.System has been deprecated and marked for removal
  * attempt to synchronize on an instance of a value-based class
* Upgrade errorprone to work with Java 17:
  * add suggested compiler args `-Xep:DeadException:WARN -Xep:GuardedBy:OFF`
  * Add additional errorprone exclusions and fix trivial errors
  * Erroprone fixes:
    * [AssertEqualsArgumentOrderChecker] Arguments are swapped in assertEquals-like call
    * [AttemptedNegativeZero] -0 is the same as 0. For the floating-point negative zero, use -0.0
    * [BigDecimalEquals] BigDecimal#equals has surprising behavior: it also compares scale.
    * [ClassCanBeStatic] Inner class is non-static but does not reference enclosing class
    * [ExtendsObject] `T extends Object` is redundant (unless you are using the Checker Framework)
    * [IdentityHashMapUsage] IdentityHashMap usage shouldn't be intermingled with Map
    * [InlineMeInliner] Migrate (via inlining) away from `Strings.repeat()`
    * [InputStreamSlowMultibyteRead] Please also override int read(byte[], int, int), otherwise multi-byte reads from this input stream are likely to be slow.
    * [IterableAndIterator] Class should not implement both `Iterable` and `Iterator` (e.g. KMLRawReader)
    * [JUnitIncompatibleType] The types of this assertion are mismatched: type `X` is not compatible with `Y`
    * [MissingCasesInEnumSwitch] Non-exhaustive switch; either add a default or handle the remaining cases
    * [ModifyCollectionInEnhancedForLoop] Modifying a collection while iterating over it in a loop may cause a ConcurrentModificationException to be thrown or lead to undefined behavior
    * [ObjectToString] X is final and does not override Object.toString, so converting it to a string will print its identity
    * [SameNameButDifferent] e.g.: The name `Geometry` refers to [org.geoserver.wps.ppio.GMLPPIO.GML2.Geometry, org.geoserver.wps.ppio.GMLPPIO.GML3.Geometry]
       within this file. It may be confusing to have the same name refer to multiple types. Consider qualifying them for clarity.
    * [SuperCallToObjectMethod] `super.hashCode()` is equivalent to `System.identityHashCode(this)` here
    * [TypeNameShadowing] Found type parameters shadowing other declared types
    * [UnnecessaryLongToIntConversion] Converting a long or Long to an int to pass as a long parameter is usually not necessary. If this
       conversion is intentional, consider `Longs.constrainToRange()` instead.
    * [UnsafeReflectiveConstructionCast] Prefer `asSubclass` instead of casting the result of `newInstance`, to detect classes
       of incorrect type before invoking their constructors.  This way, if the class is of the incorrect type,it will throw an
       exception before invoking its constructor.
    * [URLEqualsHashCode] Avoid hash-based containers of java.net.URL--the containers rely on equals() and hashCode(), which
       cause java.net.URL to make blocking internet connections.

+++ src/extension/geofence/geofence-server/src/test/java/org/geoserver/geofence/server/xstream/GeofencePersisterSerializationTest.java
@@ -69,7 +69,7 @@ public class GeofencePersisterSerializationTest {
         assertEquals("LIMIT", rule.getAccess());
         assertEquals("DE_USNG_UTM18", rule.getLayer());
         assertEquals("geonode", rule.getWorkspace());
-        assertEquals(1, rule.getPriority().intValue());
+        assertEquals(Long.valueOf(1), rule.getPriority());
 
         assertNotNull(rule.getLimits());
